+++ Apps/Examples/Examples/All Examples/DistanceExpressionExample.swift
@@ -17,6 +17,9 @@ class DistanceExpressionExample: UIViewController, ExampleProtocol {
 
         mapView.mapboxMap.onNext(.mapLoaded) { _ in
             self.addCircleLayer()
+
+            // The following line is just for testing purposes.
+            self.finish()
         }
     }
 
@@ -26,7 +29,7 @@ class DistanceExpressionExample: UIViewController, ExampleProtocol {
 
         // Create a `GeoJSONSource` from a Turf geometry.
         var source = GeoJSONSource()
-        point = Feature(geometry: .point(Point(center)))
+        point = Feature(geometry: Point(center))
 
         // Filter out POI labels that are more than 150 meters from the point.
         self.filterPoiLabels()