+++ Apps/Examples/Examples/Extensions/UIColor+Random.swift
@@ -1,0 +1,8 @@
+import Foundation
+import UIKit
+
+extension UIColor {
+    static var random: UIColor {
+        UIColor(red: .random(in: 0...1), green: .random(in: 0...1), blue: .random(in: 0...1), alpha: 1)
+    }
+}