+++ Apps/Examples/Examples/Models/ExampleProtocol.swift
@@ -17,3 +17,16 @@ extension ExampleProtocol {
         }
     }
 }
+
+extension ExampleProtocol where Self: UIViewController {
+    // Presents an alert with a given title.
+    public func showAlert(with title: String) {
+        let alertController = UIAlertController(title: title,
+                                                message: nil,
+                                                preferredStyle: .alert)
+
+        alertController.addAction(UIAlertAction(title: "Close", style: .default, handler: nil))
+
+        present(alertController, animated: true, completion: nil)
+    }
+}