+++ Apps/Examples/ExamplesTests/TestableExampleTests.swift
@@ -62,8 +62,7 @@ class TestableExampleTests: XCTestCase {
         case .completed:
             break
         case .timedOut:
-            // TODO: check if this is a failure
-            print("Example timed out, was this intentional? Call finish() if possible.")
+            XCTFail("Example: \(example.title) timed out. Don't forget to call finish().")
         default:
             XCTFail("Expectation failed with \(result)")
         }