+++ Tests/MapboxMapsTests/Foundation/Extensions/Core/FeatureExtensionValueTests.swift
@@ -15,7 +15,7 @@ final class FeatureExtensionValueTests: XCTestCase {
     func testInitializationWithWithNonNilValues() throws {
         let value = Int.random(in: 0..<100)
         let features = Array.random(withLength: .random(in: 0..<10)) { () -> Feature in
-            var feature = Feature(geometry: .point(Point(.random())))
+            var feature = Feature(geometry: Point(.random()))
             feature.identifier = .number(Double(Int.random(in: (.min)...(.max))))
             return feature
         }
@@ -38,7 +38,7 @@ final class FeatureExtensionValueTests: XCTestCase {
 
     func testNonNilFeatures() {
         let features = Array.random(withLength: .random(in: 0..<10)) { () -> Feature in
-            var feature = Feature(geometry: .point(Point(.random())))
+            var feature = Feature(geometry: Point(.random()))
             feature.identifier = .number(Double(Int.random(in: (.min)...(.max))))
             return feature
         }.map(MapboxCommon.Feature.init(_:))